Can use callbacks and plot
Call backs functional
Plot functional
Task 1 & 2 complete

+++ Functions/EncoderListener.m
@@ -3,7 +3,7 @@ function EncoderListener(~, latestMessage)
     global encoderTime;
     x= latestMessage.Vector.X;
     y= latestMessage.Vector.Y;
-    t= double(latestMessage.Header.Stamp.Sec) +(double(latestMessage.Header.Stamp.Nsec)/1000000000);
+    t= double(double(latestMessage.Header.Stamp.Sec) +(double(latestMessage.Header.Stamp.Nsec)/1000000000));
     enc = [x, y];
     
     encoderData = enc;
